+++ packages/evepilot-bootstrap/src/evepilot/bootstrap/config_apply/__init__.py
@@ -1,0 +1,14 @@
+"""Console config apply primitives."""
+
+from evepilot.bootstrap.config_apply.config_file import load_config_lines
+from evepilot.bootstrap.config_apply.models import ConfigApplyResult
+from evepilot.bootstrap.config_apply.models import ConfigCommandResult, ConfigLine
+from evepilot.bootstrap.config_apply.runner import apply_config_lines
+
+__all__ = [
+    "ConfigApplyResult",
+    "ConfigCommandResult",
+    "ConfigLine",
+    "apply_config_lines",
+    "load_config_lines",
+]