+++ packages/evepilot-eve-ng/src/evepilot/eve_ng/__init__.py
@@ -3,5 +3,18 @@
 from evepilot.eve_ng.client import EveNgClient
 from evepilot.eve_ng.errors import EveNgError
 from evepilot.eve_ng.models import EveNgNode
+from evepilot.eve_ng.schemas import EveNgNodeConsoleResult, EveNgNodeResult
+from evepilot.eve_ng.schemas import EveNgNodesResult
+from evepilot.eve_ng.service import get_node, get_node_console, list_nodes
 
-__all__ = ["EveNgClient", "EveNgError", "EveNgNode"]
+__all__ = [
+    "EveNgClient",
+    "EveNgError",
+    "EveNgNode",
+    "EveNgNodeConsoleResult",
+    "EveNgNodeResult",
+    "EveNgNodesResult",
+    "get_node",
+    "get_node_console",
+    "list_nodes",
+]